Follow up to #5825
This commit contains the follow up items from https://github.com/eclipse/omr/pull/5825

This includes:
- making `Trc_MM_MemorySubSpaceTarok_calculateExpandSize_Exit1` and `Trc_MM_MemorySubSpaceTarok_calculateTargetContractSize_Entry` obsolete
- `dnssExpectedTimeRatioMaximum` and `dnssExpectedTimeRatioMinimum` were removed, since they are no longer used. The newer version (dnssExpectedRatioMaximum/Minimum) are more flexible for different GC policies
- `heapExpansionGCTimeThreshold` and `heapContractionGCTimeThreshold` were removed, since they are no longer used. The newer versions (`heapExpansionGCRatioThreshold` and `heapContractionGCRatioThreshold`) are more flexible for different GC policies

+++ gc/stats/GlobalVLHGCStats.hpp
@@ -52,8 +52,8 @@ public:
 		uint64_t avgPgcTimeUs;
 		uint64_t avgPgcIntervalUs;
 		uint64_t pgcCountSinceGMPEnd;
-		uint64_t reservedSize;
-		uint64_t freeTenure;
+		uintptr_t reservedSize;
+		uintptr_t freeTenure;
 		intptr_t edenRegionChange;
 		bool readyToResizeAtGlobalEnd;
